requirements.txt
Django==4.1
djangorestframework==3.14.0
django-filter==22.1
django-environ==0.9.0
django-cors-headers==3.13.0
drf-spectacular==0.25.1
djoser==2.1.0
django-phonenumber-field[phonenumbers]==7.0.2
django-model-utils==4.3.1
